correcoes na api

+++ api-catalogo/src/main/java/com/catalogodeproduto/apicatalogo/dto/CorProdutoDTO.java
@@ -1,6 +1,6 @@
 package com.catalogodeproduto.apicatalogo.dto;
 
-public record CorProdutoDTO(Long id, String descricao, String imagem, Integer produtoId){
+public record CorProdutoDTO(Long id, String descricao, String imagem, Long produtoId){
 
 
 }
